+++ core/app/models/spree/shipment.rb
@@ -261,7 +261,7 @@ module Spree
     end
 
     def to_package
-      package = Stock::Package.new(stock_location)
+      package = Stock::Package.new(order, stock_location)
       package.shipment = self
       inventory_units.includes(:variant).joins(:variant).group_by(&:state).each do |state, state_inventory_units|
         package.add_multiple state_inventory_units, state.to_sym